add config for storage
